Describe How to Use Docker to Run/Install BendoBot
Add a subsection to the READMEs "Installation"-section describing the
process of running/installing the bot using Docker and name this section
"Docker".
Also move the original content of the "Installation"-section into a new
subsection called "Classic", since this content describes the classic way
of downloading and compiling the bot.

Signed-off-by: Julian Schacher <jspp@posteo.net>